@@ -898,12 +898,9 @@ AndroidBuilder.prototype.validate = function validate(logger, config, cli) {
 	}
 	cli.tiapp.properties['ti.deploytype'] = { type: 'string', value: this.deployType };
 
-	// Titanium's gradle build system currently ignores these settings.
+	// Fetch Java max heap size settings.
 	this.javacMaxMemory = cli.tiapp.properties['android.javac.maxmemory'] && cli.tiapp.properties['android.javac.maxmemory'].value || config.get('android.javac.maxMemory', '3072M');
-	this.javacSource = cli.tiapp.properties['android.javac.source'] && cli.tiapp.properties['android.javac.source'].value || config.get('android.javac.source', '1.7');
-	this.javacTarget = cli.tiapp.properties['android.javac.target'] && cli.tiapp.properties['android.javac.target'].value || config.get('android.javac.target', '1.7');
 	this.dxMaxMemory = cli.tiapp.properties['android.dx.maxmemory'] && cli.tiapp.properties['android.dx.maxmemory'].value || config.get('android.dx.maxMemory', '3072M');
-	this.dxMaxIdxNumber = cli.tiapp.properties['android.dx.maxIdxNumber'] && cli.tiapp.properties['android.dx.maxIdxNumber'].value || config.get('android.dx.maxIdxNumber', '65536');
 
 	// Transpilation details
 	this.transpile = cli.tiapp['transpile'] !== false; // Transpiling is an opt-out process now
@@ -1281,13 +1278,6 @@ AndroidBuilder.prototype.validate = function validate(logger, config, cli) {
 	// determine the abis to support
 	this.abis = this.validABIs;
 	const customABIs = cli.tiapp.android && cli.tiapp.android.abi && cli.tiapp.android.abi.indexOf('all') === -1;
-	if (!customABIs && this.deployType === 'production') {
-		// If a users has not specified the abi tag in the tiapp,
-		// remove 'x86' from production builds 'x86' devices are scarce;
-		// this is predominantly used for emulators
-		// we can save 16MB+ by removing this from release builds
-		this.abis.splice(this.abis.indexOf('x86'), 1);
-	}
 	if (customABIs) {
 		this.abis = cli.tiapp.android.abi;
 		this.abis.forEach(function (abi) {
@@ -1558,6 +1548,15 @@ AndroidBuilder.prototype.validate = function validate(logger, config, cli) {
 							process.exit(1);
 						}
 					}
+				} else {
+					// Limit application build ABI to that of provided native modules.
+					this.abis = this.abis.filter(abi => {
+						if (!module.manifest.architectures.includes(abi)) {
+							this.logger.warn(__('Module %s does not contain %s ABI. Application will build without %s ABI support!', module.id.cyan, abi.cyan, abi.cyan));
+							return false;
+						}
+						return true;
+					});
 				}
 
 				// scan the module for any CLI hooks
@@ -2259,6 +2258,27 @@ AndroidBuilder.prototype.generateRootProjectFiles = async function generateRootP
 	gradlew.logger = this.logger;
 	await gradlew.installTemplate(path.join(this.platformPath, 'templates', 'gradle'));
 
+	// Create a "gradle.properties" file. Will add network proxy settings if needed.
+	// Note: Enable Jetifier to replace all Google Support library references with AndroidX in all pre-built JARs.
+	//       This is needed because using both libraries will cause class name collisions, causing a build failure.
+	const gradleProperties = await gradlew.fetchDefaultGradleProperties();
+	gradleProperties.push({ key: 'android.useAndroidX', value: 'true' });
+	gradleProperties.push({ key: 'android.enableJetifier', value: 'true' });
+	gradleProperties.push({ key: 'org.gradle.jvmargs', value: `-Xmx${this.javacMaxMemory}` });
+	await gradlew.writeGradlePropertiesFile(gradleProperties);
+
+	// Copy optional "gradle.properties" file contents from Titainum project to the above generated file.
+	// These properties must be copied to the end of the file so that they can override Titanium's default properties.
+	const customGradlePropertiesFilePath = path.join(this.projectDir, 'platform', 'android', 'gradle.properties');
+	if (await fs.exists(customGradlePropertiesFilePath)) {
+		const targetGradlePropertiesFilePath = path.join(this.buildDir, 'gradle.properties');
+		const fileContent = await fs.readFile(customGradlePropertiesFilePath);
+		await fs.appendFile(targetGradlePropertiesFilePath,
+			'\n\n'
+			+ '# The below was copied from project file: ./platform/android/gradle.properties\n'
+			+ fileContent.toString() + '\n');
+	}
+
 	// Create a "local.properties" file providing a path to the Android SDK/NDK directories.
 	const androidNdkPath = this.androidInfo.ndk ? this.androidInfo.ndk.path : null;
 	await gradlew.writeLocalPropertiesFile(this.androidInfo.sdk.path, androidNdkPath);
@@ -2435,15 +2455,32 @@ AndroidBuilder.prototype.generateAppProject = async function generateAppProject(
 		dexerHook('', [ '', '' ], {}, resolve);
 	});
 
+	// Acquire the app's version integer code and name to be written to "build.gradle" later.
+	let versionCode = '1';
+	let versionName = this.tiapp.version ? this.tiapp.version : '1';
+	if (this.customAndroidManifest) {
+		const versionInfo = this.customAndroidManifest.getAppVersionInfo();
+		if (versionInfo) {
+			if (versionInfo.versionCode) {
+				versionCode = versionInfo.versionCode;
+			}
+			if (versionInfo.versionName) {
+				versionName = versionInfo.versionName;
+			}
+		}
+	}
+
 	// Generate a "build.gradle" file for this project from the SDK's "app.build.gradle" EJS template.
 	// Note: Google does not support setting "maxSdkVersion" via gradle script.
 	let buildGradleContent = await fs.readFile(path.join(this.templatesDir, 'app.build.gradle'));
 	buildGradleContent = ejs.render(buildGradleContent.toString(), {
 		applicationId: this.appid,
 		compileSdkVersion: this.compileSdkVersion,
+		dexJavaMaxHeapSize: this.dxMaxMemory,
 		minSdkVersion: this.minSDK,
 		targetSdkVersion: this.targetSDK,
-		versionName: this.tiapp.version ? this.tiapp.version : '1',
+		versionCode: versionCode,
+		versionName: versionName,
 		libFilePaths: this.libFilePaths,
 		libProjectNames: this.libProjectNames,
 		libDependencyStrings: this.libDependencyStrings,
@@ -2568,7 +2605,10 @@ AndroidBuilder.prototype.copyResources = function copyResources(next) {
 				let isDrawable = false;
 
 				if (m && m.length >= 4 && m[3]) {
-					const destFilename = m[3].toLowerCase();
+					const destFilename = m[3];
+					const destLowerCaseFilename = destFilename.toLowerCase();
+					const extMatch = destLowerCaseFilename.match(drawableExtRegExp);
+					const origExt = extMatch && extMatch[1] || '';
 
 					destDir = path.join(
 						_t.buildAppMainResDir,
@@ -2577,11 +2617,30 @@ AndroidBuilder.prototype.copyResources = function copyResources(next) {
 
 					if (splashScreenRegExp.test(filename)) {
 						// we have a splash screen image
-						const extMatch = destFilename.match(drawableExtRegExp);
-						const origExt = extMatch && extMatch[1] || '';
 						to = path.join(destDir, 'background' + origExt);
 					} else {
-						to = path.join(destDir, filename);
+						// We have a drawable image file. (Rename it if it contains invalid characters.)
+						let warningMessages = [];
+						if (destFilename.includes('/') || destFilename.includes('\\')) {
+							warningMessages.push(__('- Files cannot be put into subdirectories.'));
+						}
+						let destFilteredFilename = destLowerCaseFilename.replace(drawableExtRegExp, '');
+						destFilteredFilename = destFilteredFilename.replace(/[^a-z0-9_]/g, '_') + origExt;
+						if (destFilteredFilename !== destFilename) {
+							warningMessages.push(__('- Names must contain only lowercase a-z, 0-9, or underscore.'));
+						}
+						if (/^\d/.test(destFilteredFilename)) {
+							warningMessages.push(__('- Names cannot start with a number.'));
+							destFilteredFilename = '_' + destFilteredFilename;
+						}
+						if (warningMessages.length > 0) {
+							_t.logger.warn(__(`Invalid "res" file: ${path.relative(_t.projectDir, from)}`));
+							for (const nextMessage of warningMessages) {
+								_t.logger.warn(nextMessage);
+							}
+							_t.logger.warn(__(`- Titanium will rename to: ${destFilteredFilename}`));
+						}
+						to = path.join(destDir, destFilteredFilename);
 					}
 					isDrawable = true;
 				} else if (m = relPath.match(relSplashScreenRegExp)) {
@@ -3037,12 +3096,13 @@ AndroidBuilder.prototype.generateRequireIndex = async function generateRequireIn
 	// Fetch relative paths to all of the app's *.js and *.json files.
 	const filePathDictionary = {};
 	const normalizedAssetsDir = this.buildAppMainAssetsDir.replace(/\\/g, '/');
-	(function walk(dir) {
-		fs.readdirSync(dir).forEach(function (fileName) {
-			const filePath = path.join(dir, fileName);
-			const stat = fs.statSync(filePath);
+	const walkDir = async (directoryPath) => {
+		const fileNameArray = await fs.readdir(directoryPath);
+		for (const fileName of fileNameArray) {
+			const filePath = path.join(directoryPath, fileName);
+			const stat = await fs.stat(filePath);
 			if (stat.isDirectory()) {
-				walk(filePath);
+				await walkDir(filePath);
 			} else if (stat.isFile()) {
 				const lowerCaseFileName = fileName.toLowerCase();
 				if (lowerCaseFileName.endsWith('.js') || lowerCaseFileName.endsWith('.json')) {
@@ -3051,8 +3111,9 @@ AndroidBuilder.prototype.generateRequireIndex = async function generateRequireIn
 					filePathDictionary[normalizedFilePath] = 1;
 				}
 			}
-		});
-	}(this.buildAppMainAssetsResourcesDir));
+		}
+	};
+	await walkDir(this.buildAppMainAssetsResourcesDir);
 	for (const filePath of this.jsFilesToEncrypt) {
 		filePathDictionary['Resources/' + filePath.replace(/\\/g, '/')] = 1;
 	}
@@ -3108,9 +3169,11 @@ AndroidBuilder.prototype.getNativeModuleBindings = function getNativeModuleBindi
 AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() {
 	this.logger.info('Generating Java files');
 
-	const copyTemplate = (src, dest, ejsParams) => {
+	const copyTemplate = async (src, dest, ejsParams) => {
 		this.logger.debug(__('Copying template %s => %s', src.cyan, dest.cyan));
-		fs.writeFileSync(dest, ejs.render(fs.readFileSync(src).toString(), ejsParams));
+		let fileContent = await fs.readFile(src);
+		fileContent = ejs.render(fileContent.toString(), ejsParams);
+		await fs.writeFile(dest, fileContent);
 	};
 
 	// Fetch Java proxy class information from all modules.
@@ -3178,10 +3241,10 @@ AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() 
 	}
 
 	// Copy main application Java classes.
-	fs.ensureDirSync(this.buildGenAppIdDir);
-	copyTemplate(
+	await fs.ensureDir(this.buildGenAppIdDir);
+	await copyTemplate(
 		path.join(this.templatesDir, 'AppInfo.java'),
-		path.join(this.buildGenAppIdDir, this.classname + 'AppInfo.java'),
+		path.join(this.buildGenAppIdDir, `${this.classname}AppInfo.java`),
 		{
 			appid: this.appid,
 			buildType: this.buildType,
@@ -3189,9 +3252,9 @@ AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() 
 			deployType: this.deployType,
 			tiapp: this.tiapp
 		});
-	copyTemplate(
+	await copyTemplate(
 		path.join(this.templatesDir, 'App.java'),
-		path.join(this.buildGenAppIdDir, this.classname + 'Application.java'),
+		path.join(this.buildGenAppIdDir, `${this.classname}Application.java`),
 		{
 			appid: this.appid,
 			classname: this.classname,
@@ -3199,9 +3262,9 @@ AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() 
 			deployType: this.deployType,
 			encryptJS: this.encryptJS
 		});
-	copyTemplate(
+	await copyTemplate(
 		path.join(this.templatesDir, 'Activity.java'),
-		path.join(this.buildGenAppIdDir, this.classname + 'Activity.java'),
+		path.join(this.buildGenAppIdDir, `${this.classname}Activity.java`),
 		{
 			appid: this.appid,
 			classname: this.classname
@@ -3216,24 +3279,25 @@ AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() 
 	// Generate the JavaScript-based activity classes.
 	const android = this.tiapp.android;
 	if (android && android.activities) {
-		const activityTemplate = fs.readFileSync(path.join(this.templatesDir, 'JSActivity.java')).toString();
-		Object.keys(android.activities).forEach(function (name) {
-			var activity = android.activities[name];
+		const activityTemplate = (await fs.readFile(path.join(this.templatesDir, 'JSActivity.java'))).toString();
+		for (const activityName in android.activities) {
+			const activity = android.activities[activityName];
 			this.logger.debug(__('Generating activity class: %s', activity.classname.cyan));
-			fs.writeFileSync(path.join(this.buildGenAppIdDir, activity.classname + '.java'), ejs.render(activityTemplate, {
+			const fileContent = ejs.render(activityTemplate, {
 				appid: this.appid,
 				activity: activity
-			}));
-		}, this);
+			});
+			await fs.writeFile(path.join(this.buildGenAppIdDir, `${activity.classname}.java`), fileContent);
+		}
 	}
 
 	// Generate the JavaScript-based Service classes.
 	if (android && android.services) {
-		const serviceTemplate = fs.readFileSync(path.join(this.templatesDir, 'JSService.java')).toString();
-		const intervalServiceTemplate = fs.readFileSync(path.join(this.templatesDir, 'JSIntervalService.java')).toString();
-		const quickSettingsServiceTemplate = fs.readFileSync(path.join(this.templatesDir, 'JSQuickSettingsService.java')).toString();
-		Object.keys(android.services).forEach(function (name) {
-			const service = android.services[name];
+		const serviceTemplate = (await fs.readFile(path.join(this.templatesDir, 'JSService.java'))).toString();
+		const intervalServiceTemplate = (await fs.readFile(path.join(this.templatesDir, 'JSIntervalService.java'))).toString();
+		const quickSettingsServiceTemplate = (await fs.readFile(path.join(this.templatesDir, 'JSQuickSettingsService.java'))).toString();
+		for (const serviceName in android.services) {
+			const service = android.services[serviceName];
 			let tpl = serviceTemplate;
 			if (service.type === 'interval') {
 				tpl = intervalServiceTemplate;
@@ -3244,11 +3308,12 @@ AndroidBuilder.prototype.generateJavaFiles = async function generateJavaFiles() 
 			} else {
 				this.logger.debug(__('Generating service class: %s', service.classname.cyan));
 			}
-			fs.writeFileSync(path.join(this.buildGenAppIdDir, service.classname + '.java'), ejs.render(tpl, {
+			const fileContent = ejs.render(tpl, {
 				appid: this.appid,
 				service: service
-			}));
-		}, this);
+			});
+			await fs.writeFile(path.join(this.buildGenAppIdDir, `${service.classname}.java`), fileContent);
+		}
 	}
 };
 
@@ -3290,7 +3355,7 @@ AndroidBuilder.prototype.generateI18N = async function generateI18N() {
 		return locale;
 	}
 
-	Object.keys(data).forEach(function (locale) {
+	for (const locale of Object.keys(data)) {
 		const localeSuffixName = (locale === 'en' ? '' : '-' + resolveRegionName(locale));
 		const dirPath = path.join(this.buildAppMainResDir, `values${localeSuffixName}`);
 		const filePath = path.join(dirPath, 'ti_i18n_strings.xml');
@@ -3304,7 +3369,6 @@ AndroidBuilder.prototype.generateI18N = async function generateI18N() {
 		appnameNode.appendChild(dom.createTextNode(appname));
 		root.appendChild(dom.createTextNode('\n\t'));
 		root.appendChild(appnameNode);
-
 		data[locale].strings && Object.keys(data[locale].strings).forEach(function (name) {
 			if (name.indexOf(' ') !== -1) {
 				badStringNames[locale] || (badStringNames[locale] = []);
@@ -3318,13 +3382,12 @@ AndroidBuilder.prototype.generateI18N = async function generateI18N() {
 				root.appendChild(node);
 			}
 		});
-
 		root.appendChild(dom.createTextNode('\n'));
 
 		this.logger.debug(__('Writing %s strings => %s', locale.cyan, filePath.cyan));
-		fs.ensureDirSync(dirPath);
-		fs.writeFileSync(filePath, '<?xml version="1.0" encoding="UTF-8"?>\n' + dom.documentElement.toString());
-	}, this);
+		await fs.ensureDir(dirPath);
+		await fs.writeFile(filePath, '<?xml version="1.0" encoding="UTF-8"?>\n' + dom.documentElement.toString());
+	}
 
 	if (Object.keys(badStringNames).length) {
 		this.logger.error(__('Found invalid i18n string names:'));
