@@ -301,6 +301,9 @@ AndroidModuleBuilder.prototype.validate = function validate(logger, config, cli)
 				logger.warn(__('Building with Android SDK %s which hasn\'t been tested against Titanium SDK %s', ('' + this.targetSDK).cyan, this.titaniumSdkVersion));
 			}
 
+			// get javac params
+			this.javacMaxMemory = cli.timodule.properties['android.javac.maxmemory'] && cli.timodule.properties['android.javac.maxmemory'].value || config.get('android.javac.maxMemory', '3072M');
+
 			// detect java development kit
 			appc.jdk.detect(config, null, function (jdkInfo) {
 				if (!jdkInfo.version) {
@@ -534,10 +537,17 @@ AndroidModuleBuilder.prototype.generateRootProjectFiles = async function generat
 	this.logger.info(__('Generating root project files'));
 
 	// Copy our SDK's gradle files to the build directory. (Includes "gradlew" scripts and "gradle" directory tree.)
-	// The below install method will also generate a "gradle.properties" file.
 	const gradlew = new GradleWrapper(this.buildDir);
 	gradlew.logger = this.logger;
 	await gradlew.installTemplate(path.join(this.platformPath, 'templates', 'gradle'));
+
+	// Create a "gradle.properties" file. Will add network proxy settings if needed.
+	const gradleProperties = await gradlew.fetchDefaultGradleProperties();
+	gradleProperties.push({
+		key: 'org.gradle.jvmargs',
+		value: `-Xmx${this.javacMaxMemory} -Dkotlin.daemon.jvm.options="-Xmx${this.javacMaxMemory}"`
+	});
+	await gradlew.writeGradlePropertiesFile(gradleProperties);
 
 	// Create a "local.properties" file providing a path to the Android SDK/NDK directories.
 	await gradlew.writeLocalPropertiesFile(this.androidInfo.sdk.path, this.androidInfo.ndk.path);
@@ -603,6 +613,7 @@ AndroidModuleBuilder.prototype.generateModuleProject = async function generateMo
 		moduleMavenArtifactId: mavenArtifactId,
 		moduleName: this.manifest.name,
 		moduleVersion: this.manifest.version,
+		moduleArchitectures: this.manifest.architectures.split(' '),
 		tiBindingsJsonPath: path.join(this.platformPath, 'titanium.bindings.json'),
 		tiMavenUrl: encodeURI('file://' + path.join(this.platformPath, 'm2repository').replace(/\\/g, '/')),
 		tiSdkModuleTemplateDir: this.moduleTemplateDir,
